Move http classes to http package
