+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/data/CommentRepository.kt
@@ -15,6 +15,8 @@ import com.neaniesoft.vermilion.postdetails.domain.entities.CommentId
 import com.neaniesoft.vermilion.postdetails.domain.entities.ControversialIndex
 import com.neaniesoft.vermilion.postdetails.domain.entities.DurationString
 import com.neaniesoft.vermilion.postdetails.domain.entities.UpVotesCount
+import com.neaniesoft.vermilion.postdetails.data.http.CommentApiService
+import com.neaniesoft.vermilion.postdetails.data.http.CommentResponse
 import com.neaniesoft.vermilion.posts.domain.entities.AuthorName
 import com.neaniesoft.vermilion.posts.domain.entities.PostId
 import com.neaniesoft.vermilion.posts.domain.entities.Score
