+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/data/http/CommentResponse.kt
@@ -1,0 +1,9 @@
+package com.neaniesoft.vermilion.postdetails.data.http
+
+import com.fasterxml.jackson.annotation.JsonProperty
+import com.neaniesoft.vermilion.api.entities.Listing
+
+data class CommentResponse(
+    @JsonProperty("kind") val kind: String,
+    @JsonProperty("data") val data: Listing
+)
